@@ -2,13 +2,12 @@ name: battery_plus_example
 description: Demonstrates how to use the battery_plus plugin.
 
 environment:
-  sdk: '>=2.12.0 <3.0.0'
+  sdk: '>=2.18.0 <4.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  battery_plus:
-    path: ../
+  battery_plus: ^4.0.1
 
 dev_dependencies:
   flutter_driver:
