@@ -1,8 +1,8 @@
 name: battery_plus
 description: Flutter plugin for accessing information about the battery state(full, charging, discharging).
-version: 3.0.3
+version: 4.0.1
 homepage: https://plus.fluttercommunity.dev/
-repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/
+repository: https://github.com/fluttercommunity/plus_plugins/tree/main/packages/battery_plus/battery_plus
 issue_tracker: https://github.com/fluttercommunity/plus_plugins/labels/battery_plus
 
 flutter:
@@ -35,11 +35,11 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  async: ^2.9.0
+  async: ^2.10.0
   flutter_lints: ^2.0.1
-  mockito: ^5.2.0
-  plugin_platform_interface: ^2.1.3
+  mockito: ^5.4.0
+  plugin_platform_interface: ^2.1.4
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=2.11.0"
+  sdk: ">=2.18.0 <4.0.0"
+  flutter: ">=3.3.0"
