@@ -11,11 +11,11 @@ class IosDeviceInfo extends BaseDeviceInfo {
   /// IOS device info class.
   IosDeviceInfo._({
     required Map<String, dynamic> data,
-    this.name,
-    this.systemName,
-    this.systemVersion,
-    this.model,
-    this.localizedModel,
+    required this.name,
+    required this.systemName,
+    required this.systemVersion,
+    required this.model,
+    required this.localizedModel,
     this.identifierForVendor,
     required this.isPhysicalDevice,
     required this.utsname,
@@ -26,22 +26,27 @@ class IosDeviceInfo extends BaseDeviceInfo {
   /// On iOS < 16 returns user-assigned device name
   /// On iOS >= 16 returns a generic device name if project has
   /// no entitlement to get user-assigned device name.
-  /// See: https://developer.apple.com/documentation/uikit/uidevice/1620015-name
-  final String? name;
+  /// https://developer.apple.com/documentation/uikit/uidevice/1620015-name
+  final String name;
 
   /// The name of the current operating system.
-  final String? systemName;
+  /// https://developer.apple.com/documentation/uikit/uidevice/1620054-systemname
+  final String systemName;
 
   /// The current operating system version.
-  final String? systemVersion;
+  /// https://developer.apple.com/documentation/uikit/uidevice/1620043-systemversion
+  final String systemVersion;
 
   /// Device model.
-  final String? model;
+  /// https://developer.apple.com/documentation/uikit/uidevice/1620044-model
+  final String model;
 
   /// Localized name of the device model.
-  final String? localizedModel;
+  /// https://developer.apple.com/documentation/uikit/uidevice/1620029-localizedmodel
+  final String localizedModel;
 
   /// Unique UUID value identifying the current device.
+  /// https://developer.apple.com/documentation/uikit/uidevice/1620059-identifierforvendor
   final String? identifierForVendor;
 
   /// `false` if the application is running in a simulator, `true` otherwise.
@@ -71,27 +76,27 @@ class IosDeviceInfo extends BaseDeviceInfo {
 /// See http://pubs.opengroup.org/onlinepubs/7908799/xsh/sysutsname.h.html for details.
 class IosUtsname {
   const IosUtsname._({
-    this.sysname,
-    this.nodename,
-    this.release,
-    this.version,
-    this.machine,
+    required this.sysname,
+    required this.nodename,
+    required this.release,
+    required this.version,
+    required this.machine,
   });
 
   /// Operating system name.
-  final String? sysname;
+  final String sysname;
 
   /// Network node name.
-  final String? nodename;
+  final String nodename;
 
   /// Release level.
-  final String? release;
+  final String release;
 
   /// Version level.
-  final String? version;
+  final String version;
 
   /// Hardware type (e.g. 'iPhone7,1' for iPhone 6 Plus).
-  final String? machine;
+  final String machine;
 
   /// Deserializes from the map message received from [_kChannel].
   static IosUtsname _fromMap(Map<String, dynamic> map) {
