@@ -12,22 +12,26 @@ void main() {
   const channel = MethodChannel('dev.fluttercommunity.plus/package_info');
   final log = <MethodCall>[];
 
-  channel.setMockMethodCallHandler((MethodCall methodCall) async {
-    log.add(methodCall);
-    switch (methodCall.method) {
-      case 'getAll':
-        return <String, dynamic>{
-          'appName': 'package_info_example',
-          'buildNumber': '1',
-          'packageName': 'io.flutter.plugins.packageinfoexample',
-          'version': '1.0',
-          'installerStore': null,
-        };
-      default:
-        assert(false);
-        return null;
-    }
-  });
+  TestDefaultBinaryMessengerBinding.instance.defaultBinaryMessenger
+      .setMockMethodCallHandler(
+    channel,
+    (MethodCall methodCall) async {
+      log.add(methodCall);
+      switch (methodCall.method) {
+        case 'getAll':
+          return <String, dynamic>{
+            'appName': 'package_info_example',
+            'buildNumber': '1',
+            'packageName': 'io.flutter.plugins.packageinfoexample',
+            'version': '1.0',
+            'installerStore': null,
+          };
+        default:
+          assert(false);
+          return null;
+      }
+    },
+  );
 
   tearDown(() {
     log.clear();
@@ -71,51 +75,91 @@ void main() {
 
   test('equals checks for value equality', () async {
     final info1 = PackageInfo(
-        appName: 'package_info_example',
-        buildNumber: '1',
-        packageName: 'io.flutter.plugins.packageinfoexample',
-        version: '1.0',
-        buildSignature: '',
-        installerStore: null);
+      appName: 'package_info_example',
+      buildNumber: '1',
+      packageName: 'io.flutter.plugins.packageinfoexample',
+      version: '1.0',
+      buildSignature: '',
+      installerStore: null,
+    );
     final info2 = PackageInfo(
-        appName: 'package_info_example',
-        buildNumber: '1',
-        packageName: 'io.flutter.plugins.packageinfoexample',
-        version: '1.0',
-        buildSignature: '',
-        installerStore: null);
+      appName: 'package_info_example',
+      buildNumber: '1',
+      packageName: 'io.flutter.plugins.packageinfoexample',
+      version: '1.0',
+      buildSignature: '',
+      installerStore: null,
+    );
     expect(info1, info2);
   });
 
   test('hashCode checks for value equality', () async {
     final info1 = PackageInfo(
-        appName: 'package_info_example',
-        buildNumber: '1',
-        packageName: 'io.flutter.plugins.packageinfoexample',
-        version: '1.0',
-        buildSignature: '',
-        installerStore: null);
+      appName: 'package_info_example',
+      buildNumber: '1',
+      packageName: 'io.flutter.plugins.packageinfoexample',
+      version: '1.0',
+      buildSignature: '',
+      installerStore: null,
+    );
     final info2 = PackageInfo(
-        appName: 'package_info_example',
-        buildNumber: '1',
-        packageName: 'io.flutter.plugins.packageinfoexample',
-        version: '1.0',
-        buildSignature: '',
-        installerStore: null);
+      appName: 'package_info_example',
+      buildNumber: '1',
+      packageName: 'io.flutter.plugins.packageinfoexample',
+      version: '1.0',
+      buildSignature: '',
+      installerStore: null,
+    );
     expect(info1.hashCode, info2.hashCode);
   });
 
   test('toString returns a string representation', () async {
     final info = PackageInfo(
-        appName: 'package_info_example',
-        buildNumber: '1',
-        packageName: 'io.flutter.plugins.packageinfoexample',
-        version: '1.0',
-        buildSignature: '',
-        installerStore: null);
+      appName: 'package_info_example',
+      buildNumber: '1',
+      packageName: 'io.flutter.plugins.packageinfoexample',
+      version: '1.0',
+      buildSignature: '',
+      installerStore: null,
+    );
     expect(
       info.toString(),
       'PackageInfo(appName: package_info_example, buildNumber: 1, packageName: io.flutter.plugins.packageinfoexample, version: 1.0, buildSignature: , installerStore: null)',
     );
+  });
+
+  test('data returns a map representation', () async {
+    PackageInfo.setMockInitialValues(
+      appName: 'mock_package_info_example',
+      packageName: 'io.flutter.plugins.mockpackageinfoexample',
+      version: '1.1',
+      buildNumber: '2',
+      buildSignature: '',
+      installerStore: null,
+    );
+    final info1 = await PackageInfo.fromPlatform();
+    expect(info1.data, {
+      'appName': 'mock_package_info_example',
+      'packageName': 'io.flutter.plugins.mockpackageinfoexample',
+      'version': '1.1',
+      'buildNumber': '2',
+    });
+    PackageInfo.setMockInitialValues(
+      appName: 'mock_package_info_example',
+      packageName: 'io.flutter.plugins.mockpackageinfoexample',
+      version: '1.1',
+      buildNumber: '2',
+      buildSignature: 'deadbeef',
+      installerStore: 'testflight',
+    );
+    final info2 = await PackageInfo.fromPlatform();
+    expect(info2.data, {
+      'appName': 'mock_package_info_example',
+      'packageName': 'io.flutter.plugins.mockpackageinfoexample',
+      'version': '1.1',
+      'buildNumber': '2',
+      'buildSignature': 'deadbeef',
+      'installerStore': 'testflight',
+    });
   });
 }
