@@ -1,7 +1,6 @@
 // Copyright 2019 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
-
 #import "FLTSharePlusPlugin.h"
 #import "LinkPresentation/LPLinkMetadata.h"
 #import "LinkPresentation/LPMetadataProvider.h"
@@ -334,6 +333,31 @@ TopViewControllerForViewController(UIViewController *viewController) {
                   withResult:withResult];
           if (!withResult)
             result(nil);
+        } else if ([@"shareUri" isEqualToString:call.method]) {
+          NSString *uri = arguments[@"uri"];
+
+          if (uri.length == 0) {
+            result([FlutterError errorWithCode:@"error"
+                                       message:@"Non-empty uri expected"
+                                       details:nil]);
+            return;
+          }
+
+          UIViewController *rootViewController = RootViewController();
+          if (!rootViewController) {
+            result([FlutterError errorWithCode:@"error"
+                                       message:@"No root view controller found"
+                                       details:nil]);
+            return;
+          }
+          UIViewController *topViewController =
+              TopViewControllerForViewController(rootViewController);
+
+          [self shareUri:uri
+              withController:topViewController
+                    atSource:originRect
+                    toResult:result
+                  withResult:withResult];
         } else {
           result(FlutterMethodNotImplemented);
         }
@@ -396,6 +420,20 @@ TopViewControllerForViewController(UIViewController *viewController) {
   [controller presentViewController:activityViewController
                            animated:YES
                          completion:nil];
+}
+
++ (void)shareUri:(NSString *)uri
+    withController:(UIViewController *)controller
+          atSource:(CGRect)origin
+          toResult:(FlutterResult)result
+        withResult:(BOOL)withResult {
+  NSURL *data = [NSURL URLWithString:uri];
+  [self share:@[ data ]
+         withSubject:nil
+      withController:controller
+            atSource:origin
+            toResult:result
+          withResult:withResult];
 }
 
 + (void)shareText:(NSString *)shareText
